+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_constants.py
