+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_event_mappings.py
