+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_generated/_patch.py
