+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_generated/_serialization.py
