+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_generated/aio/_vendor.py
