+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_generated/models/_patch.py
