+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_generated/py.typed
