+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_messaging_shared.py
