+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_models.py
