+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_legacy/_signature_credential_policy.py
