+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_operations/__init__.py
@@ -6,14 +6,14 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._operations import EventGridClientOperationsMixin
+from ._operations import EventGridNamespaceClientOperationsMixin
 
 from ._patch import __all__ as _patch_all
 from ._patch import *  # pylint: disable=unused-wildcard-import
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-    "EventGridClientOperationsMixin",
+    "EventGridNamespaceClientOperationsMixin",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
 _patch_sdk()
