+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_patch.py
@@ -7,12 +7,26 @@ Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python
 """
 
 from typing import List
-from .._patch import EventGridSharedAccessKeyPolicy
+from azure.core.pipeline.policies import SansIOHTTPPolicy
 from azure.core.credentials import AzureKeyCredential
-from ._client import EventGridClient as ServiceClientGenerated
-from .._legacy.aio import EventGridPublisherClient
+from ._client import EventGridNamespaceClient as ServiceClientGenerated
+from ._legacy import  EventGridPublisherClient, EventGridEvent, generate_sas, SystemEventNames
 
-class EventGridNamespaceClient(ServiceClientGenerated):
+
+
+class EventGridSharedAccessKeyPolicy(SansIOHTTPPolicy):
+    def __init__(
+        self,
+        credential: "AzureKeyCredential",
+        **kwargs # pylint: disable=unused-argument
+    ) -> None:
+        super(EventGridSharedAccessKeyPolicy, self).__init__()
+        self._credential = credential
+
+    def on_request(self, request):
+        request.http_request.headers["Authorization"] = "SharedAccessKey " + self._credential.key
+
+class EventGridNamespaceClient(ServiceClientGenerated): 
     """Azure Messaging EventGrid Namespace Client.
 
     :param endpoint: The host name of the namespace, e.g.
@@ -48,4 +62,4 @@ def patch_sdk():
     """
 
 
-__all__: List[str] = ["EventGridNamespaceClient",  "EventGridPublisherClient"]  # Add all objects you want publicly available to users at this package level
+__all__: List[str] = ["EventGridNamespaceClient", "EventGridPublisherClient", "EventGridEvent", "generate_sas", "SystemEventNames"]  # Add all objects you want publicly available to users at this package level
