+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/_vendor.py
@@ -8,7 +8,7 @@
 from abc import ABC
 from typing import List, TYPE_CHECKING, cast
 
-from ._configuration import EventGridClientConfiguration
+from ._configuration import EventGridNamespaceClientConfiguration
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -29,10 +29,10 @@ def _format_url_section(template, **kwargs):
             template = "/".join(components)
 
 
-class EventGridClientMixinABC(ABC):
+class EventGridNamespaceClientMixinABC(ABC):
     """DO NOT use this class. It is for internal typing use only."""
 
     _client: "PipelineClient"
-    _config: EventGridClientConfiguration
+    _config: EventGridNamespaceClientConfiguration
     _serialize: "Serializer"
     _deserialize: "Deserializer"
