+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/aio/__init__.py
@@ -6,10 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._client import EventGridClient
-from ._version import VERSION
-
-__version__ = VERSION
+from ._client import EventGridNamespaceClient
 
 try:
     from ._patch import __all__ as _patch_all
@@ -19,7 +16,7 @@ except ImportError:
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-
+    "EventGridNamespaceClient",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
