+++ sdk/eventgrid/azure-eventgrid/azure/eventgrid/aio/_vendor.py
@@ -8,7 +8,7 @@
 from abc import ABC
 from typing import TYPE_CHECKING
 
-from ._configuration import EventGridClientConfiguration
+from ._configuration import EventGridNamespaceClientConfiguration
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -17,10 +17,10 @@ if TYPE_CHECKING:
     from .._serialization import Deserializer, Serializer
 
 
-class EventGridClientMixinABC(ABC):
+class EventGridNamespaceClientMixinABC(ABC):
     """DO NOT use this class. It is for internal typing use only."""
 
     _client: "AsyncPipelineClient"
-    _config: EventGridClientConfiguration
+    _config: EventGridNamespaceClientConfiguration
     _serialize: "Serializer"
     _deserialize: "Deserializer"
